+++ src/assets/css/main.css
@@ -1,0 +1,57 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+
+html, body, #app {
+  font-family: 'Nunito', sans-serif;
+  background-color: #f7fafc;
+  height: 100vh;
+}
+
+.tab {
+  @apply flex justify-center items-center flex-wrap;
+  gap: .75rem;
+}
+
+.tabs-wrapper {
+    flex-wrap:wrap;
+    justify-content: center;
+    gap: .75rem;
+    border-style: none
+}
+
+.tab-item {
+    border-radius: 9999px;
+    border-width: 1px;
+    --tw-border-opacity: 1;
+    border-color: rgb(209 213 219 / var(--tw-border-opacity, 1));
+    padding: .375rem 1rem;
+    font-size: .875rem;
+    line-height: 1.25rem
+}
+
+.tab-item:after {
+    content: none!important
+}
+
+.tab-item:hover {
+    --tw-border-opacity: 1;
+    border-color: rgb(101 151 255 / var(--tw-border-opacity, 1));
+    --tw-bg-opacity: 1;
+    background-color: rgb(101 151 255 / var(--tw-bg-opacity, 1));
+    --tw-text-opacity: 1;
+    color: rgb(255 255 255 / var(--tw-text-opacity, 1))
+}
+
+.tab-item.active {
+    --tw-border-opacity: 1;
+    border-color: rgb(101 177 255 / var(--tw-border-opacity, 1));
+    --tw-bg-opacity: 1;
+    background-color: rgb(101 177 255 / var(--tw-bg-opacity, 1));
+    --tw-text-opacity: 1;
+    color: rgb(255 255 255 / var(--tw-text-opacity, 1));
+    --tw-shadow: 0 1px 3px 0 rgb(0 0 0 / .1), 0 1px 2px -1px rgb(0 0 0 / .1);
+    --tw-shadow-colored: 0 1px 3px 0 var(--tw-shadow-color), 0 1px 2px -1px var(--tw-shadow-color);
+    box-shadow: var(--tw-ring-offset-shadow, 0 0 #0000),var(--tw-ring-shadow, 0 0 #0000),var(--tw-shadow)
+}
