+++ src/components/common/InfoHead.vue
@@ -1,10 +1,10 @@
 <template>
   <div class="header mb-4">
-    <h2 class="text-2xl font-bold">火车票生成器</h2>    
+    <h2 class="text-2xl font-bold">纪念车票生成器</h2>    
       <p class="text-sm text-gray-500 mt-2">
       仅供学习交流使用，不得用于商业或违法用途。
-      图标与车票版式版权归中国铁路及相关集团所有。
-      产生的任何法律后果与本项目及作者无关。
+      产生的任何法律后果与本项目及作者无关。 <br>
+      如遇字体无法加载，请尝试切换车票种类。
       </p>
   </div>
 </template>
