+++ src/main.ts
@@ -1,4 +1,5 @@
-import './assets/main.css'
+import './assets/css/main.css'
+import './assets/css/fonts.css'
 
 import { createApp } from 'vue'
 import App from './App.vue'
