+++ src/types.ts
@@ -1,16 +1,19 @@
 export type TicketData = {
-  id: string // 火车票 ID
-  redId: string // 左上角红色 ID
-  ticketOffice: string // 售票点
+  redId: string // 上票号
+  id: string // 下票号
   startStation: string // 出发地
   endStation: string // 目的地
+  checkGate?: string // 检票口
+  ticketOffice: string // 售票点
   trainNumber: string // 车次
+  price: number // 价格
   date: string // 日期
   time: string // 时间
-  price: number // 价格
+
 
   // 座位信息
-  seatType: string // 座位类型
+  seatType: string // 席别
+  berth: string // 铺位
   seatCarriage: string // 车厢号
   seatNumber: string // 座位号
 
@@ -19,8 +22,8 @@ export type TicketData = {
   passengerId: string // 身份证号
 
   // 其他信息
-  seatTypeCustom?: string // 自定义座位类型
-  checkGate?: string // 检票口
+  qrCodeId: string // 二维码内容
+  isChild?: boolean // 是否儿童票
   isStudent?: boolean // 是否学生票
   isDiscount?: boolean // 是否优惠票
